Re-namespace a few classes
